[Consensus] Introduce VoteData struct to be used by VoteMsg
Summary:
VoteMsg and QuorumCert are semantically very similar: they're referring to the same fields.
QuorumCert is carrying LedgerInfo with multiple signatures, while VoteInfo is carrying just a single signature.
Today these two structs define the same fields separately, which is unncessary duplication.
We're going to introduce a `VoteData` struct shared by both `QuorumCert` and `VoteMsg`.
In this commit we're adding it to VoteMsg only, the followup change is going to add it to QuorumCert as well.

Note that this change is purely technical and should not change any logic at all.

Testing: existing unit test coverage

Issues: ref #402