[Proof] Refactor TransactionListWithProof
Now that we have `TransactionListProof` that consists of all the
components of the proof, including `AccumulatorRangeProof` and the list
of `TransactionInfo`s, we can change `TransactionListWithProof` to
include only the list of transactions and the proof.

The complex part of proof verification has been implemented in
`AccumulatorRangeProof::verify`, so the rest of the logic becomes
relatively simple.

Closes: #1501
Approved by: msmouse