[config] Eliminate path accessors on NodeConfig
NodeConfig historically exposed accessors to paths on it directly so
that it could pull the data directory from the base config. Now we move
base config to Arc and share that during load time to the appropriate
configs.

As a follow up todo, it would be good to tighten access to the
BaseConfig so that a change in NodeConfig affects all via enforcement
which it does not now.

Closes: #1815
Approved by: mimoo