[Consensus] Fix the state_sync_on_timeout test
Summary:
The test was written prior to us introducing the logic of helping the stale participants and prior to us unifying the votes / timeouts,
hence it was fundamentally not doing the thing that it was intended to test.
With some recent unrelated changes this test started failing (being stuck). Had to just rewrite to use a slightly different flow.

Test:
the new test consistently works locally on my machine

Closes: #1872
Approved by: andll