[Executor] Add an optional epoch_change_li field to execute_and_commit_chunk
Summary:
This field carries a LedgerInfo corresponding to the end of the epoch, even if the verified target LedgerInfo is somewhere in the future (as it's the case with the waypoints).
In addition, minor cleanup of the existing code to use the ProcessedVMOutput wrapper functions instead of lower-level checks.
Note that the new optional field is not used yet, so this change is a noop at this point.

Testing:
The existing functionality is tested with the existing unit tests.
The new functionality (epoch change LedgerInfo) is not tested yet: didn't find a simple way to generate a test reconfiguration transaction.

Issues: ref #1384

Closes: #1966
Approved by: dmitri-perelman