[language] reject zero-sized struct declarations in bytecode verifier
Zero-sized structs present problems for serialization and deserialization. They are sometimes convenient for the programmer, but not expressive given that you can always use a dummy field.

This PR tweaks the bytecode verifier to reject structs with zero fields. The only use of this pattern in production code was LibraCoin.MintCapability. I added a dummy boolean field to let this continue working.

Almost all of the changes here are silly ones in tests that used zero-sized structs.

Closes: #1999
Approved by: tnowacki